requirements.txt
beautifulsoup4==4.5.0
celery>=3.1
certifi
freezegun
html5lib
libfaketime
lxml==3.6.4
mock
nose-timer
nose==1.3.1
pandas>=0.18
parameterized
python-dateutil==2.2
pytz
requests-cache
requests-mock
requests==2.9.1
Sphinx==1.2.2
xlrd